refactored and editor screen size minimized
