+++ src/components/generic/editable-field-components/index.tsx
@@ -1,0 +1,5 @@
+export { default as MetaformTextFieldComponent } from "./metaform-text-field-component";
+export { default as MetaformHtmlComponent } from "./metaform-html-field-component";
+export { default as MetaformRadioFieldComponent } from "./metaform-radio-field-component";
+export { default as MetaformSubmitFieldComponent } from "./metaform-submit-field-component";
+export { default as MetaformNumberFieldComponent } from "./metaform-radio-field-component";
