+++ src/components/generic/editable-field-components/metaform-radio-field-component.tsx
@@ -1,12 +1,13 @@
-import { FormControl, FormControlLabel, FormLabel, Radio, RadioGroup } from "@material-ui/core";
+import { FormControl, FormControlLabel, FormLabel, Radio, RadioGroup, WithStyles, withStyles } from "@material-ui/core";
 import React from "react";
 import { MetaformField, MetaformFieldOption } from "../../../generated/client";
 import strings from "../../../localization/strings";
+import styles from "../../../styles/generics/editable-field-components/metaform-radio-field-component";
 
 /**
  * Component props
  */
-interface Props {
+interface Props extends WithStyles<typeof styles> {
   field?: MetaformField;
   fieldId?: string;
   fieldLabelId?: string;
@@ -113,5 +114,6 @@ export class MetaformRadioFieldComponent extends React.Component<Props, State> {
       />
     )
   }
-
 }
+
+export default withStyles(styles)(MetaformRadioFieldComponent)
