+++ src/styles/generics/editable-field-components/metaform-number-field-component.ts
@@ -1,0 +1,5 @@
+import { createStyles } from "@material-ui/core";
+
+export default createStyles({
+
+});
