+++ src/styles/generics/editable-field-components/metaform-submit-field-component.ts
@@ -1,0 +1,6 @@
+
+import { createStyles } from "@material-ui/core";
+
+export default createStyles({
+
+});
